Get unit auras from memory
Less dependencies on function hooks, and safer code. I am not sure however, whether it is more performant now cause of marshalling and limitations of .NET Framework.
Memory utilities of .NET Core would help tremendously, but we've got to work with what we have.

Also decommisioning function hook of CGUnit_C::GetAura as it is no longer needed.