Hiding hooks from Warden
Before scanning, hooks from targeted area are removed and after scan they are restored.

For now hardcoded protection for only unsafe hook that we do - InvalidPtrCheck, but
extending protection for generic hooks or memory writes should be trivial.

For now, protection works ONLY when dll is injected BEFORE login.
If you injected after, in order to turn on protection, you have to re-login.