Rework remoting architecture
Remoting server is in injected dll and client is host app.
Reworked services such as LEL and IM to work with the new architecture.
The reason is it will provide greater stability and possiblity to reattach to WoW client, without reinjection

todo:
Debug print from the remote is now not functional.
Rework it the same way as the services.